requirements.txt
altair==5.5.0
astroid==3.3.9
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dill==0.4.0
gitdb==4.0.12
gitpython==3.1.44
idna==3.10
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markupsafe==3.0.2
mccabe==0.7.0
narwhals==1.36.0
nodeenv==1.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
pyright==1.1.399
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requirements-parser==0.11.0
rpds-py==0.24.0
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
spotipy==2.25.1
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
types-setuptools==79.0.0.20250422
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
